DAOS-9231 test: fix dm_negative_space_dcp
Move to negative space test to hw small since VM performance
is not reliable.

Skip-unit-tests: true
Skip-fault-injection-test: true
Test-repeat: 15
Test-tag: dm_negative_space_dcp
